モデルの２重管理を解消
不必要なログを削除
